# swagger_ui/openapi_ms1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment Data Service API",
    "description": "Documentação completa das rotas do microserviço investment_data_service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/admin/clientes/id/{cliente_id}": {
      "get": {
        "summary": "Buscar cliente por ID",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "nome": { "type": "string" },
                    "cpf": { "type": "string" },
                    "email": { "type": "string" },
                    "telefone": { "type": "string" },
                    "correntista": { "type": "integer" },
                    "score_credito": { "type": "number" },
                    "saldo_cc": { "type": "number" },
                    "admin": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/investidor/cliente/{cliente_id}": {
      "get": {
        "summary": "Buscar cliente + investidor pelo ID do cliente",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investidor encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "nome": { "type": "string" },
                    "cpf": { "type": "string" },
                    "email": { "type": "string" },
                    "telefone": { "type": "string" },
                    "correntista": { "type": "integer" },
                    "score_credito": { "type": "number" },
                    "saldo_cc": { "type": "number" },
                    "admin": { "type": "integer" },
                    "perfil_investidor": { "type": "string" },
                    "patrimonio_total": { "type": "number" },
                    "data_cadastro": { "type": "string", "format": "date" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Investidor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/investidor/id/{investidor_id}": {
      "get": {
        "summary": "Buscar investidor por ID do cliente",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "investidor_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investidor encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_cliente": { "type": "string" },
                    "perfil_investidor": { "type": "string" },
                    "patrimonio_total": { "type": "number" },
                    "data_cadastro": { "type": "string", "format": "date" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Investidor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/clientes/id/{id}": {
      "patch": {
        "summary": "Atualizar dados de um cliente",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string" },
                  "cpf": { "type": "string" },
                  "email": { "type": "string" },
                  "telefone": { "type": "string" },
                  "correntista": { "type": "integer" },
                  "score_credito": { "type": "number" },
                  "saldo_cc": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "400": {
            "description": "CPF, email ou telefone em uso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "erro": { "type": "string" } }
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/admin/investidor/id/atualiza/{id_cliente}": {
      "patch": {
        "summary": "Atualizar dados de investidor",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "perfil_investidor": { "type": "string" },
                  "patrimonio_total": { "type": "number" },
                  "data_cadastro": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investidor atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cliente não é investidor ou erro ao atualizar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "erro": { "type": "string" } }
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/admin/investimentos/atualiza/{id_investimento}": {
      "patch": {
        "summary": "Atualizar investimento",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": { "type": "string" },
                  "tipo_investimento": { "type": "string" },
                  "valor_investido": { "type": "number" },
                  "data_aplicacao": { "type": "string" },
                  "rentabilidade": { "type": "number" },
                  "ativo": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investimento atualizado"
          },
          "400": {
            "description": "Erro ao atualizar"
          },
          "404": {
            "description": "Investimento não encontrado"
          }
        }
      }
    },
...
    "/admin/transacoes/{id_cliente}": {
      "get": {
        "summary": "Listar transações de um cliente",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transações encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_transacao": { "type": "integer" },
                      "cliente_id": { "type": "string" },
                      "tipo": { "type": "string" },
                      "valor": { "type": "number" },
                      "data": { "type": "string" },
                      "descricao": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/clientes/delete/id/{id}": {
      "delete": {
        "summary": "Deletar cliente pelo ID",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente deletado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sucesso": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/investimento/id/{id_investimento}": {
      "get": {
        "summary": "Buscar investimento por ID",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investimento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_investimento": { "type": "integer" },
                    "cliente_id": { "type": "string" },
                    "tipo_investimento": { "type": "string" },
                    "valor_investido": { "type": "number" },
                    "data_aplicacao": { "type": "string" },
                    "rentabilidade": { "type": "number" },
                    "ativo": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Investimento não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },  
...
}